fix: Implement clone and checkout on TC run, copy on local run [NMO-504]
Quick hack to solve the TeamCity server "error: object directory ...
does not exist; check .git/objects/info/alternates". The idea is to
clone when the script is run on a TC server as we know the code was
committed to a branch and copy when when locally as the code might not
be committed yet.

[skip release]